requirements.txt
absl-py==0.12.0
matplotlib==3.4.2
more-itertools==4.2.0
numpy==1.20.3
ortools==9.0.9048
pandas==1.2.4
protobuf==3.17.1
six==1.14.0
flask
